pyproject.toml
[project]
name = "falldetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "load-dotenv>=0.1.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "ultralytics>=8.3.217",
]
